auth with cookie

+++ src/app/api/assignees/route.ts
@@ -3,6 +3,11 @@ import { NextRequest, NextResponse } from "next/server";
 
 
 export async function POST(req: NextRequest) {
+  if(req.headers.get('coookie')?.slice(6,)) return NextResponse.json({
+    success: false,
+    message: 'Unauthorization'
+  })
+
   const {taskId, userId} = await req.json();
   
   await prisma.assignees.create({
@@ -18,6 +23,10 @@ export async function POST(req: NextRequest) {
 }
 
 export async function DELETE(req: NextRequest) {
+  if(req.headers.get('coookie')?.slice(6,)) return NextResponse.json({
+    success: false,
+    message: 'Unauthorization'
+  })
   const {taskId, userId} = await req.json();
 
   await prisma.assignees.deleteMany({
